Integrate CommentsApi within room.ts
