+++ packages/liveblocks-core/src/comments/CommentsApiError.ts
@@ -1,0 +1,11 @@
+import type { JsonObject } from "../lib/Json";
+
+export class CommentsApiError extends Error {
+  constructor(
+    public message: string,
+    public status: number,
+    public details?: JsonObject
+  ) {
+    super(message);
+  }
+}
